Combine mock ObjectFactory, ObjectStore, and SearchWrapper.
* Combine the mock SearchWrapper in TraverserFactoryFixture with
  SearchMock. This removed both uses of Capture.

* Move MockObjectStore unchanged into FileNetProxies in place of
  ObjectStoreMock.

* Remove the mostly unimplemented and unused MockObjectFactory.

Remove two tests from DocumentTraverserTest:

* testGetDocumentList_initialCheckpoint was a duplicate of
  testGetDocIds_noResults, once the queries were always smoke
  tested by SearchMock, thus removing the need for capture.

* testNullObjectSet_nullDocuments was a vestige of a
  FileDocumentList test in v3 that passed a null to the
  constructor. We don't have a prod class that takes an object
  set, and the NPE was now thrown by the mocks, which is silly.
